+++ compiler-core/src/erlang.rs
@@ -6,6 +6,7 @@ mod pattern;
 #[cfg(test)]
 mod tests;
 
+use crate::build::Target;
 use crate::type_::is_prelude_module;
 use crate::{
     ast::{CustomType, Function, Import, ModuleConstant, TypeAlias, *},
@@ -260,8 +261,11 @@ fn register_imports(
             public,
             name,
             arguments: args,
+            implementations,
             ..
-        }) if *public || overridden_publicity.contains(name) => {
+        }) if implementations.supports(Target::Erlang)
+            && (*public || overridden_publicity.contains(name)) =>
+        {
             exports.push(atom_string(name.to_string()).append("/").append(args.len()))
         }
 
@@ -375,6 +379,12 @@ fn module_function<'a>(
     // Private external functions don't need to render anything, the underlying
     // Erlang implementation is used directly at the call site.
     if function.external_erlang.is_some() && !function.public {
+        return None;
+    }
+
+    // If the function has no suitable Erlang implementation then there is nothing
+    // to generate for it.
+    if !function.implementations.supports(Target::Erlang) {
         return None;
     }