+++ compiler-core/src/erlang/tests/external_fn.rs
@@ -265,3 +265,17 @@ pub fn do() -> Int
 "#
     );
 }
+
+#[test]
+fn javascript_only() {
+    assert_erl!(
+        r#"
+pub fn should_be_generated(x: Int) -> Int {
+  x
+}
+
+@external(javascript, "one", "one")
+pub fn should_not_be_generated(x: Int) -> Int
+"#
+    );
+}