+++ compiler-core/src/javascript.rs
@@ -5,6 +5,8 @@ mod pattern;
 mod tests;
 mod typescript;
 
+use crate::analyse::TargetSupport;
+use crate::build::Target;
 use crate::type_::PRELUDE_MODULE_NAME;
 use crate::{
     ast::{CustomType, Function, Import, ModuleConstant, TypeAlias, *},
@@ -38,10 +40,15 @@ pub struct Generator<'a> {
     tracker: UsageTracker,
     module_scope: im::HashMap<EcoString, usize>,
     current_module_name_segments_count: usize,
+    target_support: TargetSupport,
 }
 
 impl<'a> Generator<'a> {
-    pub fn new(line_numbers: &'a LineNumbers, module: &'a TypedModule) -> Self {
+    pub fn new(
+        line_numbers: &'a LineNumbers,
+        module: &'a TypedModule,
+        target_support: TargetSupport,
+    ) -> Self {
         let current_module_name_segments_count = module.name.split('/').count();
 
         Self {
@@ -50,6 +57,7 @@ impl<'a> Generator<'a> {
             module,
             tracker: UsageTracker::default(),
             module_scope: Default::default(),
+            target_support,
         }
     }
 
@@ -169,33 +177,38 @@ impl<'a> Generator<'a> {
         imports.register_module(path, [], [member]);
     }
 
-    pub fn statement(&mut self, statement: &'a TypedDefinition) -> Vec<Output<'a>> {
+    pub fn statement(&mut self, statement: &'a TypedDefinition) -> Option<Output<'a>> {
         match statement {
-            Definition::TypeAlias(TypeAlias { .. }) => vec![],
+            Definition::TypeAlias(TypeAlias { .. }) => None,
 
             // Handled in collect_imports
-            Definition::Import(Import { .. }) => vec![],
+            Definition::Import(Import { .. }) => None,
 
             // Handled in collect_definitions
-            Definition::CustomType(CustomType { .. }) => vec![],
+            Definition::CustomType(CustomType { .. }) => None,
 
             Definition::ModuleConstant(ModuleConstant {
                 public,
                 name,
                 value,
                 ..
-            }) => vec![self.module_constant(*public, name, value)],
+            }) => Some(self.module_constant(*public, name, value)),
 
-            Definition::Function(Function {
-                arguments,
-                name,
-                body,
-                public,
-                external_javascript: None,
-                ..
-            }) => vec![self.module_function(*public, name, arguments, body)],
+            Definition::Function(function) => {
+                // If there's an external JavaScript implementation then it will be imported,
+                // so we don't need to generate a function definition.
+                if function.external_javascript.is_some() {
+                    return None;
+                }
 
-            Definition::Function(_) => vec![],
+                // If the function does not support JavaScript then we don't need to generate
+                // a function definition.
+                if !function.implementations.supports(Target::JavaScript) {
+                    return None;
+                }
+
+                self.module_function(function)
+            }
         }
     }
 
@@ -425,40 +438,51 @@ impl<'a> Generator<'a> {
         let _ = self.module_scope.insert(name.into(), 0);
     }
 
-    fn module_function(
-        &mut self,
-        public: bool,
-        name: &'a EcoString,
-        args: &'a [TypedArg],
-        body: &'a [TypedStatement],
-    ) -> Output<'a> {
-        let argument_names = args
+    fn module_function(&mut self, function: &'a TypedFunction) -> Option<Output<'a>> {
+        let argument_names = function
+            .arguments
             .iter()
             .map(|arg| arg.names.get_variable_name())
             .collect();
         let mut generator = expression::Generator::new(
             self.module.name.clone(),
             self.line_numbers,
-            name.clone(),
+            function.name.clone(),
             argument_names,
             &mut self.tracker,
             self.module_scope.clone(),
         );
-        let head = if public {
+        let head = if function.public {
             "export function "
         } else {
             "function "
         };
-        let body = generator.function_body(body, args)?;
-        Ok(docvec![
+
+        let body = match generator.function_body(&function.body, function.arguments.as_slice()) {
+            // No error, let's continue!
+            Ok(body) => body,
+
+            // There is an error coming from some expression that is not supported on JavaScript
+            // and the target support is not enforced. In this case we do not error, instead
+            // returning nothing which will cause no function to be generated.
+            Err(error) if error.is_unsupported() && !self.target_support.is_enforced() => {
+                return None
+            }
+
+            // Some other error case which will be returned to the user.
+            Err(error) => return Some(Err(error)),
+        };
+
+        let document = docvec![
             head,
-            maybe_escape_identifier_doc(name),
-            fun_args(args, generator.tail_recursion_used),
+            maybe_escape_identifier_doc(function.name.as_str()),
+            fun_args(function.arguments.as_slice(), generator.tail_recursion_used),
             " {",
             docvec![line(), body].nest(INDENT).group(),
             line(),
             "}",
-        ])
+        ];
+        Some(Ok(document))
     }
 
     fn register_module_definitions_in_scope(&mut self) {
@@ -486,8 +510,9 @@ pub fn module(
     line_numbers: &LineNumbers,
     path: &Utf8Path,
     src: &EcoString,
+    target_support: TargetSupport,
 ) -> Result<String, crate::Error> {
-    let document = Generator::new(line_numbers, module)
+    let document = Generator::new(line_numbers, module, target_support)
         .compile()
         .map_err(|error| crate::Error::JavaScript {
             path: path.to_path_buf(),
@@ -515,6 +540,16 @@ pub fn ts_declaration(
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Error {
     Unsupported { feature: String, location: SrcSpan },
+}
+
+impl Error {
+    /// Returns `true` if the error is [`Unsupported`].
+    ///
+    /// [`Unsupported`]: Error::Unsupported
+    #[must_use]
+    pub fn is_unsupported(&self) -> bool {
+        matches!(self, Self::Unsupported { .. })
+    }
 }
 
 fn fun_args(args: &'_ [TypedArg], tail_recursion_used: bool) -> Document<'_> {