+++ compiler-core/src/javascript/tests/externals.rs
@@ -258,3 +258,31 @@ pub fn one(x: Int) -> Int {
 "#
     );
 }
+
+#[test]
+fn erlang_only() {
+    assert_js!(
+        r#"
+pub fn should_be_generated(x: Int) -> Int {
+  x
+}
+
+@external(erlang, "one", "one")
+pub fn should_not_be_generated(x: Int) -> Int
+"#
+    );
+}
+
+#[test]
+fn erlang_bit_patterns() {
+    assert_js!(
+        r#"
+pub fn should_not_be_generated(x) {
+  case x {
+    <<_, rest:bits>> -> rest
+    _ -> x
+  }
+}
+"#
+    );
+}